+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [0.1.3] - 2025-09-25
+- Updated simplejson=3.19.2
+- Added support for Python 3.13
+- Updated mock==5.2.0
+- Updated python-lib to v4.1.0
+- Updated rabbitmq-lib to v2.4.2
+
+
 ## [0.1.2] - 2025-05-02
 - Added option (-T) to view queue message counts.
 - Added ability to check threshold queue message count (-j) option.